requirements.txt
amqp==2.6.0
arrow==0.16.0
asgiref==3.2.10
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.3
django-3-jet==1.0.8
django-braces==1.14.0
django-celery-beat==2.0.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-markdown-deux==1.0.5
django-oauth-toolkit==1.3.0
django-phonenumber-field==4.0.0
django-recurrence==1.10.3
django-rest-framework-social-oauth2==1.1.0
django-scheduler==0.9.3
django-timezone-field==4.0
django-tinymce==3.2.0
django-wysiwyg==0.8.0
djangorestframework==3.11.0
gunicorn==20.0.4
icalendar==4.0.6
idna==2.10
install==1.3.3
kombu==4.6.11
markdown2==2.3.9
model-bakery==1.1.0
oauthlib==3.1.0
phonenumbers==8.12.6
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.2.0
